Refactor so that all Nodes think of their source in terms of character index by default (instead of lines, as we were doing before).
This way is more uniform: nodes that start at the beginning of a line just have a start_column of 0. We're going to need good ways to access things smaller than a line as we get into tiny nodes, so we might as well build character-level access in as the default, and make lines an abstraction on top if we need them.